Cleanup patient hompage code by refactoring Video Card hidden state
